Qualcomm AI Engine Direct - Adapt to new IR capture flow
Summary:
- Change existent IR capture flow (exir.capture) to torch.export.export
- Add custom decomposition table for mitigating maintaining effort
- Fix breakages encountered and make sure all tests passed as well

+++ backends/qualcomm/passes/i64_to_i32.py
@@ -28,7 +28,7 @@ class I64toI32(ExportPass):
             )
         else:
             if meta_val.dtype == torch.int64:
-                node.meta["val"] = meta_val.to(torch.int32)
+                node.meta["val"] = meta_val.to(torch.float)
 
     def _cast_to_int32(self, graph_module: torch.fx.GraphModule):
         for n in graph_module.graph.nodes:
